+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/api/oid4vci/CredentialEnvelopeDto.java
@@ -9,6 +9,7 @@ package ch.admin.bj.swiyu.issuer.api.oid4vci;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import org.springframework.http.HttpStatus;
 
 /**
  * Helper DTO providing the http content type alongside the credential
@@ -19,4 +20,5 @@ import lombok.Getter;
 public class CredentialEnvelopeDto {
     private String contentType;
     private String oid4vciCredentialJson;
+    private HttpStatus httpStatus;
 }