+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Expanded the credential endpoint to accept the new credential-endpoint (with corresponding response)
+  defined [here](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#name-credential-endpoint),
+  and the deferred credential endpoint which is
+  defined [here](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#name-deferred-credential-endpoin).
+  These endpoints can be used by setting the custom header `SWIYU-API-Version=2`. These endpoints are not yet pentested.
 - Expanded cnf to contain correct structure while still providing the old one. Example:
 
 ```json